Refactored package structure and cleaned up namespaces
Standardized package declarations across core modules to improve codebase consistency and maintainability.
