+++ app/src/main/java/com/example/bluetooth_chat/data/chat/BluetoothDeviceMapper.kt
@@ -1,8 +1,8 @@
-package com.plcoding.bluetoothchat.data.chat
+package com.example.bluetooth_chat.data.chat
 
 import android.annotation.SuppressLint
 import android.bluetooth.BluetoothDevice
-import com.plcoding.bluetoothchat.domain.chat.BluetoothDeviceDomain
+import com.example.bluetooth_chat.domain.chat.BluetoothDeviceDomain
 
 @SuppressLint("MissingPermission")
 fun BluetoothDevice.toBluetoothDeviceDomain(): BluetoothDeviceDomain {
