+++ app/src/main/java/com/example/bluetooth_chat/domain/chat/TransferFailedException.kt
@@ -1,4 +1,4 @@
-package com.plcoding.bluetoothchat.domain.chat
+package com.example.bluetooth_chat.domain.chat
 
 import java.io.IOException
