+++ app/src/main/java/com/example/bluetooth_chat/ui/screens/BluetoothDevicesScreen.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.plcoding.bluetoothchat.domain.chat.BluetoothDevice
-import com.plcoding.bluetoothchat.presentation.BluetoothUiState
+import com.example.bluetooth_chat.domain.chat.BluetoothDevice
+import com.example.bluetooth_chat.presentation.BluetoothUiState
 
 @Composable
 fun BluetoothDevicesScreen(
